updated index.js to run getBlogPosts and then run updateBlogPosts. Now all I have to do is run "node ." in this directory to sync my posts directory with the blog posts I have on Dev.to

+++ getBlogPosts.js
@@ -3,6 +3,7 @@ const axios = require("axios");
 const fs = require("fs");
 const util = require("util");
 const writeFile = util.promisify(fs.writeFile);
+const exists = util.promisify(fs.exists);
 
 const getBlogPosts = async () => {
   try {
@@ -11,8 +12,13 @@ const getBlogPosts = async () => {
     );
     response.data.forEach(async el => {
       let post = await axios.get(`https://dev.to/api/articles/${el.id}`);
-      await writeFile(`posts/${post.data.slug}.md`, post.data.body_markdown);
-      console.log(`successfully saved ${post.data.title} posts/${post.data.slug}.md`);
+      let fileExist = await exists(`posts/${post.data.slug}.md`)
+      if (!fileExist) {
+        await writeFile(`posts/${post.data.slug}.md`, post.data.body_markdown);
+        console.log(`successfully saved ${post.data.title} to posts/${post.data.slug}.md`);
+      } else {
+        console.log(`posts/${post.data.slug}.md already exists. moving on...`)
+      }
     });
   } catch (e) {
     console.log(e);
